working
working

+++ Controller/cycoController.js
@@ -54,10 +54,18 @@ const register = asyncHandler(async (req, res) => {
         const existingUser = await CycoModel.findOne({ email, isVerified: false });
 
 
-        if (existingUser && existingUser.otp) {
-            res.status(200).send('OTP already sent recently. Please check your email.');
-            return;
+
+        if (existingUser) {
+            throw new Error("user with this email already exit as psychologist")
         }
+
+
+
+
+        // if (existingUser && existingUser.otp) {
+        //     res.status(200).send('OTP already sent recently. Please check your email.');
+        //     return;
+        // }
 
         const otp = generateOTP();
 
@@ -287,6 +295,13 @@ const login = asyncHandler(async (req, res) => {
             return;
         }
 
+        // Compare entered password with the password in the database
+        if (password !== user.password) {
+            res.status(401).json({ message: 'Invalid password' });
+            return;
+        }
+
+
         if (user.isVerified !== true) {
             setTimeout(async () => {
                 await CycoModel.deleteMany({ email: user?.email });
@@ -298,12 +313,7 @@ const login = asyncHandler(async (req, res) => {
         }
 
 
-        // Compare entered password with the password in the database
-        if (password !== user.password) {
-            res.status(401).json({ message: 'Invalid password' });
-            return;
-        }
-
+        
 
 
         const mailOptions = {
